fix:去除告警
git-svn-id: http://192.168.20.11/sequoiadb/trunk@26343 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/spt/sptBsonobjArray.cpp
@@ -186,7 +186,7 @@ namespace engine
             detail = BSON( SPT_ERR << "The 1st param must be INT32 value" ) ;
             goto error ;
          }
-         if ( idValue < _vecObj.size() )
+         if ( idValue < (INT32)_vecObj.size() )
          {
             /// set the return value
             rval.getReturnVal().setValue( _vecObj[idValue] ) ;
